+++ starter/app/src/main/java/com/example/android/politicalpreparedness/data/local/database/ElectionDao.kt
@@ -2,10 +2,6 @@ package com.example.android.politicalpreparedness.data.local.database
 
 import androidx.room.*
 import com.example.android.politicalpreparedness.data.network.models.Election
-import com.example.android.politicalpreparedness.data.network.models.RepresentativeResponse
-import com.example.android.politicalpreparedness.data.network.models.VoterInfoResponse
-import retrofit2.http.GET
-import retrofit2.http.Path
 
 @Dao
 interface ElectionDao {
@@ -17,10 +13,10 @@ interface ElectionDao {
     fun getAllElections(): List<Election>
 
     @Query("SELECT * from election_table WHERE id = :id")
-    fun getElectionById(id: Int): Election
+    fun getElectionById(id: Long): Election
 
     @Query("DELETE from election_table WHERE id = :id")
-    fun deleteElection(id: Int)
+    fun deleteElection(id: Long)
 
     @Query("DELETE from election_table")
     fun clear()
